setup for 'docker-compose build'

+++ virtuoso/scripts/setup.sh
@@ -2,7 +2,7 @@
 SETTINGS_DIR=/settings
 mkdir -p $SETTINGS_DIR
 
-cd /data
+cd /data_
 
 mkdir -p dumps
 
@@ -40,7 +40,7 @@ then
     cd backups
     virtuoso-t +restore-backup $BACKUP_PREFIX +configfile /data_/virtuoso.ini
     if [ $? -eq 0 ]; then
-        cd /data
+        cd /data_
         echo "`date +%Y-%m-%dT%H:%M:%S%:z`" > .backup_restored
     else
         exit -1
@@ -75,15 +75,15 @@ then
     echo "`date +%Y-%m-%dT%H:%M:%S%:z`" > .data_loaded
 fi
 
-crudini --set virtuoso.ini HTTPServer ServerPort ${VIRT_HTTPServer_ServerPort:-$original_port}
-
 # Enable Federated
-touch /enable-federated-query.sql
-echo "grant select on \"DB.DBA.SPARQL_SINV_2\" to \"SPARQL\";" >> /enable-federated-query.sql
-echo "grant execute on \"DB.DBA.SPARQL_SINV_IMP\" to \"SPARQL\";" >> /enable-federated-query.sql
-virtuoso-t +wait && isql-v -U dba -P dba < /enable-federated-query.sql
-kill "$(ps aux | grep '[v]irtuoso-t' | awk '{print $2}')"
-echo "`date +%Y-%m-%dT%H:%M:%S%:z`" >  .enable_federated
+if [ ! -f ".enable_federated" ];
+then
+  touch /enable-federated-query.sql
+  echo "grant select on \"DB.DBA.SPARQL_SINV_2\" to \"SPARQL\";" >> /enable-federated-query.sql
+  echo "grant execute on \"DB.DBA.SPARQL_SINV_IMP\" to \"SPARQL\";" >> /enable-federated-query.sql
+  virtuoso-t +wait && isql-v -U dba -P dba < /enable-federated-query.sql
+  kill "$(ps aux | grep '[v]irtuoso-t' | awk '{print $2}')"
+  echo "`date +%Y-%m-%dT%H:%M:%S%:z`" >  .enable_federated
+fi
 
-
-exec virtuoso-t +wait +foreground
+crudini --set virtuoso.ini HTTPServer ServerPort ${VIRT_HTTPServer_ServerPort:-$original_port}
